# public/static/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CyberGuard AI - Enterprise API Documentation",
    "description": "API endpoints for the CyberGuard AI Threat Detection Platform. Enforces tenant logical isolation, Role-Based Access Control, and Swagger documentation standards.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Register a new user account with tenant partitioning",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password", "confirm_password"],
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" },
                  "confirm_password": { "type": "string" },
                  "role": { "type": "string", "enum": ["admin", "analyst", "user"], "default": "analyst" },
                  "tenant_name": { "type": "string", "default": "Default Org" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User registered successfully" },
          "400": { "description": "Validation error" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Authenticate user and return JWT token",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Token generated successfully" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
...
    "/api/threat-intel/cves": {
      "get": {
        "summary": "Search CVE database",
        "parameters": [
          { "name": "q", "in": "query", "schema": { "type": "string" } },
          { "name": "severity", "in": "query", "schema": { "type": "string" } },
          { "name": "min_cvss", "in": "query", "schema": { "type": "number", "format": "float" } },
          { "name": "min_epss", "in": "query", "schema": { "type": "number", "format": "float" } },
          { "name": "sort_by", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
    "/api/threat-intel/domain/{domain}": {
      "get": {
        "summary": "Retrieve domain reputation status",
        "parameters": [
          { "name": "domain", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
...
  }
}